Refactor to match Aeron's layout
